Move aux functions to a better location
