+++ app/actions/ats_client.py
@@ -4,12 +4,6 @@ import pydantic
 import xmltodict
 
 from datetime import datetime
-from app.actions.configurations import (
-    AuthenticateConfig,
-    PullObservationsConfig
-)
-from app.services.errors import ConfigurationNotFound
-from app.services.utils import find_config_for_action
 from xml.parsers.expat import ExpatError
 from typing import List, Optional
 
@@ -97,34 +91,6 @@ class ATSBadXMLException(Exception):
         self.message = message
         self.error = error
         super().__init__(f"'{self.status_code}: {self.message}, Error: {self.error}'")
-
-
-def get_auth_config(integration):
-    # Look for the login credentials, needed for any action
-    auth_config = find_config_for_action(
-        configurations=integration.configurations,
-        action_id="auth"
-    )
-    if not auth_config:
-        raise ConfigurationNotFound(
-            f"Authentication settings for integration {str(integration.id)} "
-            f"are missing. Please fix the integration setup in the portal."
-        )
-    return AuthenticateConfig.parse_obj(auth_config.data)
-
-
-def get_pull_config(integration):
-    # Look for the login credentials, needed for any action
-    pull_config = find_config_for_action(
-        configurations=integration.configurations,
-        action_id="pull_observations"
-    )
-    if not pull_config:
-        raise ConfigurationNotFound(
-            f"Authentication settings for integration {str(integration.id)} "
-            f"are missing. Please fix the integration setup in the portal."
-        )
-    return PullObservationsConfig.parse_obj(pull_config.data)
 
 
 def closest_transmission(transmissions, test_date):
