+++ app/actions/configurations.py
@@ -1,5 +1,8 @@
 from .core import AuthActionConfiguration, PullActionConfiguration, ExecutableActionMixin
-import pydantic  
+import pydantic
+
+from ..services.errors import ConfigurationNotFound
+from ..services.utils import find_config_for_action
 
 
 class AuthenticateConfig(AuthActionConfiguration, ExecutableActionMixin):
@@ -15,3 +18,31 @@ class PullObservationsConfig(PullActionConfiguration):
 class ProcessObservationsConfig(PullActionConfiguration):
     mortality_event_type: str = "mortality_event"
     observations_per_request: int = 200
+
+
+def get_auth_config(integration):
+    # Look for the login credentials, needed for any action
+    auth_config = find_config_for_action(
+        configurations=integration.configurations,
+        action_id="auth"
+    )
+    if not auth_config:
+        raise ConfigurationNotFound(
+            f"Authentication settings for integration {str(integration.id)} "
+            f"are missing. Please fix the integration setup in the portal."
+        )
+    return AuthenticateConfig.parse_obj(auth_config.data)
+
+
+def get_pull_config(integration):
+    # Look for the login credentials, needed for any action
+    pull_config = find_config_for_action(
+        configurations=integration.configurations,
+        action_id="pull_observations"
+    )
+    if not pull_config:
+        raise ConfigurationNotFound(
+            f"Authentication settings for integration {str(integration.id)} "
+            f"are missing. Please fix the integration setup in the portal."
+        )
+    return PullObservationsConfig.parse_obj(pull_config.data)
